Создает контроллер для отрисовки страницы с фильмами
